refactor(product): 상품 이미지 관련 코드 리펙토링
- 각 계층의 역할을 구분하도록 코드 리펙토링
- ApiResponse 적용 및 예외 처리 코드 추가

#30 #31
